core_logic

Module agents

this is where all agent structs are defined and their function implementation derive Serialize, Deserialize when needing to load data and make a JSON, TOML, YAML… derive Debug when developing to be able to pritn using {:?}, Recommended to keep in Production for logs. tracing, assertions, panic messages… derive Clone if having types that need that to be able to use .clone() like String and Vec derive Copy for copy types (on stack) like String, i32… need to have Clone as well and can then not use .clone() and reuse same field, no & derive Eq for full total equality useful for sets, maps. also floats can’t be totally equal like f32 or other floats as flaoting numbers can differ a bit (not fully total precision) derive PartialEq for use of == and != for those struct field derive Default to initialize struct with initial values so no need implementation of .new() for the struct

Structs

Agent

this is me creating a generic agent with all fields needed to make any type of agent

CallApiResponseFormat

Function

we will need to implement here as there can be several functions details added

FunctionDetails

FunctionParametersContainer

TOOLS structs and impls: we will try to match what the API is eecting receive when calling LLM on Cerebras

LlmResponse

here we will be deserializing the llm’s response, but also serialize to save it (need then to add it to all nested other structs above)

MessageHistory

this will be the buffer history of messages stored and sent to an llm, so we need to limit it a certain way changed my mind, using here just a VecDeque as we can have with_capacity() and push_back()/push_front() methods

MessageToAppend

this is the message to send after a tool call have been identified in the response, so llm have choosen a tool, we need to append to messages and send it to the llm again, and get the response and append it to the messages until tool is not called in a loop way with or without the tool_call_id: [{“content”: “Hello!”, “role”: “user”, “tool_call_id”: “…”}]``

MessagesSent

ModelSettings

we define for the agent and then maybe pick what we need from it after its definition or just use it directly need to test the api and adapt

Payload

this will be creating a dynamic payload with or without tools, with or without structout we do not need to create the fields of the struct, we just impl a function to it as we will not store any paylaod state but just use those when needed and build when needed

ResponseChoices

ResponseFormat

ResponseMessage

RoutedMessage

thisis the message transmitter sent through the channels

Schema

this is the schema of the structured output structure generic to all different schema needed in the app

SchemaFieldDetails

this is the schema of the structured output structure generic to all different schema needed in the app

StateReportPrToMain

StateReportSreToPr

StructOut

this is the structured output field and will use of the the other sturct Schema to build different structured output

TasksIdentified

this will be saving the state of indentifed tasks to be done

ToolCall

ToolFunction

Tools

Enums

AgentRole

ChoiceTool

MessageRole

this one enum is for the formating or match on messages roles

SchemaFieldType

TaskCompletion

UserType

here need to derive Hash, Eq, PartialEq because we are using this custom types in HashMap. so needed for loop,push,indexing..etc…

Functions

json_object

json_schema