core_logic::agents

Struct StructOut

pub struct StructOut {
    pub HumanRequestAnalyzerStructOut: Schema,
    pub MainAgentStructOut: Schema,
    pub PrAgentStructOut: Schema,
    pub Sre1StructOut: Schema,
    pub Sre2StructOut: Schema,
}

this is the structured output field and will use of the the other sturct Schema to build different structured output

Fields
HumanRequestAnalyzerStructOut: Schema

Schema is constructed and different ones will be stored in thos state StructOut

MainAgentStructOut: SchemaPrAgentStructOut: SchemaSre1StructOut: SchemaSre2StructOut: Schema
Implementations

impl StructOut

pub fn new( human_schema: &Schema, main_agent_schema: &Schema, pr_agent_schema: &Schema, sre1_schema: &Schema, sre2_schema: &Schema, ) -> StructOut

We construct out state StructOut

pub fn build_schema( schema_field_dict: &HashMap<String, &SchemaFieldType>, ) -> Schema

we provide a HashMap with key value and will use those to construct the schema

pub fn get_full_struct_as_map(&self) -> HashMap<String, &Schema>

this set of two functions as_map() and sturct-_out_to_json_map()
will get the output consummable version without the struct name but just its Json Value from serde

pub fn struct_out_to_json_map(struct_out: &StructOut) -> HashMap<String, Value>

function to get the full StructOut this is how to call this set of functions to have it has a dict

let json_map = StructOut::struct_out_to_json_map(&schema_big_state);
match serde_json::to_string_pretty(&json_map) {
  Ok(final_json) => println!("jsonyfied StructOut: {}", final_json),
  Err(e) => eprintln!("Error serializing schema_big_state to JSON: {}", e),
}

pub fn get_by_role(&self, role: &AgentRole) -> Option<&Schema>

Call it like that when wanting to get the schema

if let Some(schema) = agent.StructuredOutput.get_by_role(&agent.Role) {
  println!("Schema for this agent role: {:#?}", schema);
}