core_logic

Module constants

Here we will have all the constant that will be built and created those will have same lifetime as the application and would be used, specially the fields of those in order to have all needed to make dynamic API calls there will be constant engines needed for the app to work:

agents identities

modelsettings definition

Functions

agent_git_commit_work_tool

agent_git_merge_work_tool

agent_git_pull_work_tool

agent_no_tool

agent_read_file_tool

agent_write_file_tool

all_schemas_structout_constant

git_commit_work_tool

git_commit_work_tool This tool commits after work is done with writing changes to manifest file

git_merge_work_tool

git_merge_work_tool This tool git merge the work of a specific agent

git_pull_work_tool

git_pull__work_tool This tool git pull the work of a specific agent

human_schema

main_agent_merge

main_agent_merge_response_format_part

main_agent_own_task_merge_schema

main_agent_own_task_select_agent_schema

main_agent_read

main_agent_read_response_format_part

main_agent_report

main_agent_report_response_format_part

main_agent_report_to_main_agent_schema

main_merge_agent_prompt

main_merge_agent_schema

main_merge_model_message_formatted_hashmap_prompt

main_merge_model_settings

main_merge_payload_tool

main_read_agent_prompt

main_read_agent_schema

main_read_model_message_formatted_hashmap_prompt

main_read_model_settings

main_read_payload_tool

main_report_agent_prompt

main_report_agent_schema

main_report_model_message_formatted_hashmap_prompt

main_report_model_settings

main_report_payload_no_tool

model_llama3_3_70b

model_llama4_scout_17b

model_message_formatted_hashmap_prompt

model_qwen3_32b

pr_agent_own_task_pull_schema

pr_agent_own_task_select_agent_schema

pr_agent_pull

pr_agent_pull_response_format_part

pr_agent_read

pr_agent_read_response_format_part

pr_agent_report

pr_agent_report_response_format_part

pr_agent_report_to_main_agent_schema

pr_pull_agent_prompt

pr_pull_agent_schema

pr_pull_model_message_formatted_hashmap_prompt

pr_pull_model_settings

pr_pull_payload_tool

pr_read_agent_prompt

pr_read_agent_schema

pr_read_model_message_formatted_hashmap_prompt

pr_read_model_settings

pr_read_payload_tool

pr_report_agent_prompt

pr_report_agent_schema

pr_report_model_message_formatted_hashmap_prompt

pr_report_model_settings

pr_report_payload_no_tool

read_file_tool

read_file_tool This tool reads files by providing the full content of the file to be analyzed

request_analyzer_agent

request_analyzer_agent_prompt

request_analyzer_agent_schema

request_analyzer_model_settings

request_analyzer_payload_no_tool

request_analyzer_payload_tool

request_analyzer_response_format_part

sre1_agent_commit

sre1_agent_commit_response_format_part

sre1_agent_read

sre1_agent_read_response_format_part

sre1_agent_report

sre1_agent_report_response_format_part

sre1_agent_write

sre1_agent_write_response_format_part

sre1_commit_agent_prompt

sre1_commit_agent_schema

sre1_commit_model_message_formatted_hashmap_prompt

sre1_commit_model_settings

sre1_commit_payload_tool

sre1_own_task_commit_schema

sre1_own_task_read_files_schema

sre1_own_task_write_files_schema

sre1_read_agent_prompt

sre1_read_agent_schema

sre1_read_model_message_formatted_hashmap_prompt

** Sre1 Agent ModelSettings

sre1_read_model_settings

sre1_read_payload_tool

sre1_report_agent_prompt

sre1_report_agent_schema

sre1_report_model_message_formatted_hashmap_prompt

sre1_report_model_settings

sre1_report_payload_no_tool

sre1_report_to_pr_schema

sre1_write_agent_prompt

sre1_write_agent_schema

sre1_write_model_message_formatted_hashmap_prompt

sre1_write_model_settings

sre1_write_payload_tool

sre2_agent_commit

sre2_agent_commit_response_format_part

sre2_agent_read

sre2_agent_read_response_format_part

sre2_agent_report

sre2_agent_report_response_format_part

sre2_agent_write

sre2_agent_write_response_format_part

sre2_commit_agent_prompt

sre2_commit_agent_schema

sre2_commit_model_message_formatted_hashmap_prompt

sre2_commit_model_settings

sre2_commit_payload_tool

sre2_own_task_commit_schema

sre2_own_task_read_files_schema

sre2_own_task_write_files_schema

sre2_read_agent_prompt

sre2_read_agent_schema

sre2_read_model_message_formatted_hashmap_prompt

** Sre2 Agent ModelSettings

sre2_read_model_settings

sre2_read_payload_tool

sre2_report_agent_prompt

sre2_report_agent_schema

sre2_report_model_message_formatted_hashmap_prompt

sre2_report_model_settings

sre2_report_payload_no_tool

sre2_report_to_pr_schema

sre2_write_agent_prompt

sre2_write_agent_schema

sre2_write_model_message_formatted_hashmap_prompt

sre2_write_model_settings

sre2_write_payload_tool

write_file_tool

write_file_tool This tool writes files when provided with full content to be written in the manifest