core_logic

Module machine

machines contructors that will be using all our implemented struct methods engines that will be using those machines constructor or/and methods implemented in structs to build bigger picture objects this is to make it easy to call any part of the app needed. we need to create more custom errors so that it makes life easy to know at which part does the error has been triggers in our machines/engines flow

Functions

create_agent_engine

create_model_settings_engine

create_payload_engine

create_schemas_engine

create_tool_engine

engine_prompt

execute_tools_machine

This function receives the name and arguments of the tool call and dispatches the appropriate logic. The return type is a JSON value that will be added to the message history.

get_prompt_user_and_content_engine

get_specific_agent_schema_engine

machine_agent

machine_api_call

machine_api_response

this function checks on the response to see if there is any tool call used in the loop api call function

machine_create_payload_with_or_without_tools_structout

machine_final_answer

this one will return the response when there is no more tools to call

machine_history_update

machine_prompt

messages_format_engine

response_format_part_of_payload_engine

structure_final_output_from_raw_engine

tool_loop_until_final_answer_engine

tool_or_not_loop_api_call_engine