core_logic

Module merge

this is to pull work it will git checkout and git pullthe work of the correct agent using these commands format: git checkout PR_Feature_AgentX git pull agentX agentX_feature –no-edit`

Functions

main_branch

Env Vars For the agent branches to be pulled simulating pull request and the main branch for the last merge also added the upstream repo name so that it is easy to pull using that name

main_repo_path

merge_work

sre1_branch_main

sre2_branch_main