core_logic

Module prompts

All prompts for our different agents. Those are only the initial prompts tasks which might be reused several times. we will use format!() macro, when calling this prompt content and make the content dynamic this will be feeding the content argument of engine_prompt()

pub struct Agent {
  pub role: AgentRole,
  // content of message to be red by other agents  about task
  pub communication_message: HashMap<String, String>,
  pub prompt: MessagesSent,
  /// Eg. for Human request Analyzer Agent {HumanStructuredOutput.Agent: HumanStructuredOutput.Task }
  /// But at least we are free to add any key pairs
  /// use "StructOut::get_by_role(&self, role: &AgentRole)" to get it
  pub structured_output: StructOut,
  pub task_state: TaskCompletion,
  /// this is where all tools will be set and hold all necessary fields
  /// but still will need to use those fields to construct what the API will consume at the end,
 /// so we might implement a fucntion here that will for example transform enums in `String`
 pub llm: ModelSettings,
 /* ** Might need to add a field like a checklist so that agent know what need to be done next,
        Optional field so we have only the main agent with that. to keep track of work. so will need to call api also to organize work  ** */
}

Functions

human_request_agent_prompt

human_request_agent_prompt_for_structured_output

main_agent_merge_prompt

main_agent_read_and_select_agent_prompt

main_agent The main_agent will read the report from pr_agent and know which agent to merge the work from and then create a report for human sent through discord,

main_agent_report_prompt

pr_agent_pull_prompt

pr_agent_read_and_select_agent_prompt

pr_agent

pr_agent_report_prompt

sre1_agent_commit_prompt

sre1_agent_read_prompt

sre1_agent

sre1_agent_report_prompt

sre1_agent_write_prompt

sre2_agent_commit_prompt

sre2_agent_read_prompt

sre2_agent

sre2_agent_report_prompt

sre2_agent_write_prompt