core_logic

Module schemas

here will be all the different schemas like we do in Python we will use the create_schemas_engine() and inject the different agents HashMap<String, &SchemaFieldType::String>

Functions

get_schema_fields

human_request_agent_schema

human request agentschemas this one will just get tasks and affect to the right agent task requirements from this schema response will need to be saved in state so that all agents can have access to it

main_agent_merge_schema

main_agent_own_task_select_agent_schema

main_agentschemas

main_agent_report_schema

pr_agent_own_task_pull_schema

pr_agent_own_task_select_agent_schema

pr_agent_report_to_main_agent_schema

sre1_agent_own_task_commit_schema

sre1_agent_own_task_read_files_schema

sre1_agent_own_task_write_files_schema

sre1_agent_to_pr_agent_schema

sre2_agent_own_task_commit_schema

sre2_agent_own_task_read_files_schema

sre2_agent_own_task_write_files_schema

sre2_agent_to_pr_agent_schema

sre2_agentschemas